Run all checks, when check names are not specified
Previously, check names were mandatory arguments.
Now we detect when no check names were passed, and run all checks.
This makes it slightly easier to maintain the integration test we will write.
