+++ lib/checker.rb
@@ -17,5 +17,6 @@ require 'import/publishing_api_data_presenter'
 require 'import/publishing_api_importer'
 
 require 'whitelist/whitelist'
+require 'checks/report/report'
 
 require 'check_runner'
